refactor rlp decode get size

+++ lib/machine/stack/head/struct.h
@@ -8,7 +8,7 @@
 STRUCT(evm_stack)
 {
     evm_stack_l_p sl;
-    int count;
+    uint64_t count;
 };
 
 evm_stack_t stack_init();
